Lower refresh rate at night
Every ten minutes between midnight and 06:00, let's see how that works out.
Also, made it update the text timestamp, even if it didn't actually change, if the current date is different from when the text was set. This is so the weekday will be correct. I can imagine early morning "Wait, isn't it Thursday?!" situations. XD